Remove retired/unused rules & methods
